scrapePage moved to its own file
