+++ apps/playwright-scraper/src/strapi.ts
@@ -76,70 +76,66 @@ const GET_WEBPAGE_SUMMARY_BY_URL_AND_MODEL_QUERY = `
   }
 `;
 
-export const upsertWebPageSummaries = async (summaries: WebPageSummary[]) => {
+export const upsertWebPageSummaries = async (summary: WebPageSummary) => {
   try {
-    await Promise.all(
-      summaries.map(async (summary) => {
-        const data = {
-          Title: summary.title,
-          Url: summary.url,
-          LargeLanguageModel: "gpt-3.5-turbo",
-          OriginalContent: summary.content,
-          GeneratedSummary: summary.summary,
-          GeneratedKeywords: JSON.stringify(summary.keywords),
-        };
-        const updateData = {
-          Title: summary.title,
-          OriginalContent: summary.content,
-          GeneratedSummary: summary.summary,
-          GeneratedKeywords: JSON.stringify(summary.keywords),
-        };
+    const data = {
+      Title: summary.title,
+      Url: summary.url,
+      LargeLanguageModel: "gpt-3.5-turbo",
+      OriginalContent: summary.content,
+      GeneratedSummary: summary.summary,
+      GeneratedKeywords: JSON.stringify(summary.keywords),
+    };
+    const updateData = {
+      Title: summary.title,
+      OriginalContent: summary.content,
+      GeneratedSummary: summary.summary,
+      GeneratedKeywords: JSON.stringify(summary.keywords),
+    };
 
-        // Check if the WebPageSummary already exists
-        const existingSummaryResponse = await client.request<{
-          webPageSummaries: WebPageSummaryEntityResponseCollection;
-        }>(GET_WEBPAGE_SUMMARY_BY_URL_AND_MODEL_QUERY, {
-          url: summary.url,
-          model: "gpt-3.5-turbo",
+    // Check if the WebPageSummary already exists
+    const existingSummaryResponse = await client.request<{
+      webPageSummaries: WebPageSummaryEntityResponseCollection;
+    }>(GET_WEBPAGE_SUMMARY_BY_URL_AND_MODEL_QUERY, {
+      url: summary.url,
+      model: "gpt-3.5-turbo",
+    });
+
+    console.log("existingSummaryResponse: ", existingSummaryResponse);
+
+    if (
+      existingSummaryResponse.webPageSummaries.data &&
+      existingSummaryResponse.webPageSummaries.data.length > 0
+    ) {
+      const existingSummaryId =
+        existingSummaryResponse.webPageSummaries.data[0].id;
+      console.log("existingSummaryId: ", existingSummaryId);
+
+      // If it exists, update it
+      if (existingSummaryId) {
+        await client.request<{
+          updateWebPageSummary: WebPageSummaryEntityResponse;
+        }>(UPDATE_WEBPAGE_SUMMARY_MUTATION, {
+          id: existingSummaryId,
+          data: updateData,
         });
-
-        console.log("existingSummaryResponse: ", existingSummaryResponse);
-
-        if (
-          existingSummaryResponse.webPageSummaries.data &&
-          existingSummaryResponse.webPageSummaries.data.length > 0
-        ) {
-          const existingSummaryId =
-            existingSummaryResponse.webPageSummaries.data[0].id;
-          console.log("existingSummaryId: ", existingSummaryId);
-
-          // If it exists, update it
-          if (existingSummaryId) {
-            await client.request<{
-              updateWebPageSummary: WebPageSummaryEntityResponse;
-            }>(UPDATE_WEBPAGE_SUMMARY_MUTATION, {
-              id: existingSummaryId,
-              data: updateData,
-            });
-            console.log(
-              "Successfully updated WebPageSummary with ID:",
-              existingSummaryId
-            );
-          }
-        } else {
-          // If it doesn't exist, create it
-          const response = await client.request<{
-            createWebPageSummary: WebPageSummaryEntityResponse;
-          }>(CREATE_WEBPAGE_SUMMARY_MUTATION, {
-            data,
-          });
-          console.log(
-            "Successfully created WebPageSummary with ID:",
-            response.createWebPageSummary.data?.id
-          );
-        }
-      })
-    );
+        console.log(
+          "Successfully updated WebPageSummary with ID:",
+          existingSummaryId
+        );
+      }
+    } else {
+      // If it doesn't exist, create it
+      const response = await client.request<{
+        createWebPageSummary: WebPageSummaryEntityResponse;
+      }>(CREATE_WEBPAGE_SUMMARY_MUTATION, {
+        data,
+      });
+      console.log(
+        "Successfully created WebPageSummary with ID:",
+        response.createWebPageSummary.data?.id
+      );
+    }
   } catch (error) {
     console.error("Failed to upsert WebPageSummary", error);
   }
